rename controller, add class level annotations
