--- src/main/java/org/launchcode/hellospring/controllers/HelloSpringController.java
@@ -1,39 +1,0 @@
-package org.launchcode.hellospring.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-/**
- * Created by Chris Bay
- */
-@Controller
-@ResponseBody
-@RequestMapping("hello")
-public class HelloSpringController {
-
-    // Responds to /hello?name=LaunchCode
-    @RequestMapping(method = {RequestMethod.GET, RequestMethod.POST})
-    public String hello(@RequestParam String name) {
-        return "Hello, " + name + "!";
-    }
-
-    // Responds to /hello/LaunchCode
-    @GetMapping("{name}")
-    public String helloAgain(@PathVariable String name) {
-        return "Hello, " + name + "!";
-    }
-
-    @GetMapping("form")
-    public String helloForm() {
-        String html = "<html>" +
-                "<body>" +
-                "<form method = 'post' action = '/hello'>" +
-                "<input type = 'text' name = 'name' />" +
-                "<input type = 'submit' value = 'Greet Me!' />" +
-                "</form>" +
-                "</body>" +
-                "</html>";
-        return html;
-    }
-
-}
